Reimplement Marker Editor
